% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_functions.R
\name{convertToPandocSessionIdentifier}
\alias{convertToPandocSessionIdentifier}
\title{Converte título para identificador de sessão de acordo com padrão pandoc}
\usage{
convertToPandocSessionIdentifier(nome)
}
\description{
Converte título para identificador de sessão de acordo com padrão pandoc
}
