% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{escapaExpressaoRegular}
\alias{escapaExpressaoRegular}
\title{Escapa uma expressão regular para facilitar a busca pela string literal}
\usage{
escapaExpressaoRegular(texto)
}
\description{
Escapa uma expressão regular para facilitar a busca pela string literal
}
