% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_functions.R
\name{escapeRegexp}
\alias{escapeRegexp}
\title{Escapa uma expressão regular para facilitar a busca pela string literal}
\usage{
escapeRegexp(texto)
}
\description{
Escapa uma expressão regular para facilitar a busca pela string literal
}
