% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_functions.R
\name{extractFileName}
\alias{extractFileName}
\title{Extrai nome do arquivo de um caminho completo
removendo todo o texto até a última barra (/)}
\usage{
extractFileName(filename)
}
\description{
Extrai nome do arquivo de um caminho completo
removendo todo o texto até a última barra (/)
}
