% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_functions.R
\name{extractLastCommitFromGitRepository}
\alias{extractLastCommitFromGitRepository}
\title{Extract hash and datetime from last repository commit in a given directory}
\usage{
extractLastCommitFromGitRepository(directory = ".")
}
\description{
Extract hash and datetime from last repository commit in a given directory
}
