% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pom_functions.R
\name{extractProjectVersionFromPomFile}
\alias{extractProjectVersionFromPomFile}
\title{Extract project version from a pom.xml file (maven)}
\usage{
extractProjectVersionFromPomFile(arquivo)
}
\description{
Extract project version from a pom.xml file (maven)
}
