% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_functions.R
\name{extraiUltimoCommit}
\alias{extraiUltimoCommit}
\title{Extrai hash e data do último commit do repositório do diretório indicado}
\usage{
extraiUltimoCommit(diretorio = ".")
}
\description{
Extrai hash e data do último commit do repositório do diretório indicado
}
