% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{gerarDocumentos}
\alias{gerarDocumentos}
\title{Deprecated. Use renderAll}
\usage{
gerarDocumentos(arquivoPandoc, profile = "complete", pasta_destino = "build")
}
\description{
Deprecated. Use renderAll
}
