% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{gerarDocumentosLeves}
\alias{gerarDocumentosLeves}
\title{Deprecated. Use renderAll}
\usage{
gerarDocumentosLeves(arquivo, profile = "complete", pasta_destino = "build")
}
\description{
Deprecated. Use renderAll
}
