% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{identificadorSessao}
\alias{identificadorSessao}
\title{Converte título para identificador de sessão de acordo com padrão pandoc}
\usage{
identificadorSessao(nome)
}
\description{
Converte título para identificador de sessão de acordo com padrão pandoc
}
