% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{imprimirComReferencias}
\alias{imprimirComReferencias}
\title{Imprime estrutura gerada por montaDocumento incluíndo backlinks}
\usage{
imprimirComReferencias(retorno, ignoraReferenciasDe)
}
\description{
Imprime estrutura gerada por montaDocumento incluíndo backlinks
}
