% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_functions.R
\name{installUpdatedBaldessaR}
\alias{installUpdatedBaldessaR}
\title{Compila package a partir do diretório de trabalho pai do atual}
\usage{
installUpdatedBaldessaR()
}
\description{
Compila package a partir do diretório de trabalho pai do atual
}
