% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{nomeArquivo}
\alias{nomeArquivo}
\title{Extrai nome do arquivo de um caminho completo
removendo todo o texto até a última barra (/)}
\usage{
nomeArquivo(filename)
}
\description{
Extrai nome do arquivo de um caminho completo
removendo todo o texto até a última barra (/)
}
