% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_building_functions.R
\name{printWithBacklinks}
\alias{printWithBacklinks}
\title{Imprime estrutura gerada por montaDocumento incluíndo backlinks}
\usage{
printWithBacklinks(retorno, ignoraReferenciasDe)
}
\description{
Imprime estrutura gerada por montaDocumento incluíndo backlinks
}
