% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_building_functions.R
\name{readFileLines}
\alias{readFileLines}
\title{Chama função readLines, mas dispara erro informativo caso o arquivo não exista}
\usage{
readFileLines(filename)
}
\description{
Chama função readLines, mas dispara erro informativo caso o arquivo não exista
}
