% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_building_functions.R
\name{replaceSessionLinks}
\alias{replaceSessionLinks}
\title{substituí links para arquivos que foram incluídos por referências para as sessões do documento}
\usage{
replaceSessionLinks(texto, sessoes)
}
\description{
substituí links para arquivos que foram incluídos por referências para as sessões do documento
}
