% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{substituiLinksParaSessoes}
\alias{substituiLinksParaSessoes}
\title{substituí links para arquivos que foram incluídos por referências para as sessões do documento}
\usage{
substituiLinksParaSessoes(texto, sessoes)
}
\description{
substituí links para arquivos que foram incluídos por referências para as sessões do documento
}
