% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{extraiUltimoCommit}
\alias{extraiUltimoCommit}
\title{Deprecated. Use extractLastCommitFromGitRepository}
\usage{
extraiUltimoCommit(diretorio = ".")
}
\description{
Deprecated. Use extractLastCommitFromGitRepository
}
